TagPro/TagPro Replay/requirements.txt
Pillow==3.4.2
pyglet==1.2.4